partially completed addremove functionality
