+++ src/index.html
@@ -17,13 +17,6 @@
       <input type="text" name="input" id="input" placeholder="Add to your List" >
     </div>
     <ul class="items"></ul>
-    <template class="itemTemplate">
-      <li class="task">
-          <input class="input" type="checkbox" name="checkbox" id="one">
-          <p class="item-description"></p>
-        <i class="fa-solid fa-trash"></i>
-      </li>
-    </template>
     <h3>clear All Completed</h3>
   </div>
 </body>
