+++ src/modules/get.js
@@ -1,0 +1,11 @@
+/** @format */
+
+function getItems() {
+  const value = localStorage.getItem('todos') || '[]';
+
+  return JSON.parse(value);
+}
+
+const items = getItems();
+
+export default items;
