+++ src/modules/sort.js
@@ -1,0 +1,8 @@
+/** @format */
+import items from './get.js';
+
+function sort() {
+  items.sort((a, b) => a.index - b.index);
+}
+
+export default sort;
